ex 4.16

+++ exercises/chap4/ex4_16/Item.h
@@ -1,6 +1,6 @@
 #ifndef ITEM_H
 #define ITEM_H
 
-typedef float Item;
+typedef int Item;
 
 #endif
